+++ packages/checkbox/src/Checkbox.tsx
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { createDataProp } from '@leafygreen-ui/lib';
+import { createDataProp, IdAllocator } from '@leafygreen-ui/lib';
 import { css, cx } from '@leafygreen-ui/emotion';
 import { colors } from '@leafygreen-ui/theme';
 import {
@@ -26,8 +26,11 @@ export const Variant = {
 
 export type Variant = typeof Variant[keyof typeof Variant];
 
-const wrapperStyle = css`
+const wrapperStyleAnimated = css`
   transition: 300ms opacity ease-in-out;
+`;
+
+const wrapperStyle = css`
   height: ${height}px;
   width: ${height}px;
   display: inline-block;
@@ -70,8 +73,11 @@ const wrapperStyleChecked = css`
   opacity: 1;
 `;
 
-const checkboxStyleChecked = css`
+const checkboxStyleAnimated = css`
   transition: 500ms transform steps(29);
+`;
+
+const checkboxStyleChecked = css`
   transform: translate3d(${-width + height}px, 0, 0);
 `;
 
@@ -120,7 +126,7 @@ const disabledTextStyle = css`
   color: ${colors.gray[5]};
 `;
 
-interface CheckboxProps {
+interface CheckboxProps extends React.InputHTMLAttributes<HTMLInputElement> {
   variant: Variant;
   checked?: boolean;
   label: React.ReactNode;
@@ -129,11 +135,10 @@ interface CheckboxProps {
   className: string;
   onChange: React.ChangeEventHandler<HTMLInputElement>;
   bold: boolean;
+  animate?: boolean;
 }
 
-export default class Checkbox extends PureComponent<
-  CheckboxProps & React.InputHTMLAttributes<HTMLInputElement>
-> {
+export default class Checkbox extends PureComponent<CheckboxProps> {
   static displayName = 'Checkbox';
 
   static propTypes = {
@@ -146,6 +151,7 @@ export default class Checkbox extends PureComponent<
     onChange: PropTypes.func,
     bold: PropTypes.bool,
     id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    animate: PropTypes.bool,
   };
 
   static defaultProps = {
@@ -173,7 +179,17 @@ export default class Checkbox extends PureComponent<
     }
   }
 
-  checkboxId = `checkbox-${Math.floor(Math.random() * 10000000)}`;
+  private static idAllocator = IdAllocator.create('checkbox');
+  private _defaultCheckboxId?: string;
+
+  private get defaultCheckboxId(): string {
+    if (!this._defaultCheckboxId) {
+      this._defaultCheckboxId = Checkbox.idAllocator.generate();
+    }
+
+    return this._defaultCheckboxId;
+  }
+
   inputRef = React.createRef<HTMLInputElement>();
 
   onClick = (
@@ -206,12 +222,13 @@ export default class Checkbox extends PureComponent<
   };
 
   render() {
-    const checkboxId = this.props.id || this.checkboxId;
+    const checkboxId = this.props.id || this.defaultCheckboxId;
     const labelId = `${checkboxId}-label`;
 
     const {
       name = checkboxId,
       checked = this.state.checked,
+      animate = true,
       className,
       style,
       label,
@@ -305,11 +322,13 @@ export default class Checkbox extends PureComponent<
           {...checkboxWrapper.prop}
           className={cx(wrapperStyle, {
             [wrapperStyleChecked]: checked && !indeterminate && !disabled,
+            [wrapperStyleAnimated]: animate && !indeterminate && !disabled,
           })}
         >
           <div
             className={cx(checkboxStyle, checkboxBackgroundImage, {
               [checkboxStyleChecked]: checked && !indeterminate && !disabled,
+              [checkboxStyleAnimated]: animate && !indeterminate && !disabled,
             })}
           />
         </div>