+++ packages/tokens/src/spacing.ts
@@ -1,0 +1,11 @@
+const spacing = {
+  1: '4px',
+  2: '8px',
+  3: '16px',
+  4: '24px',
+  5: '32px',
+  6: '64px',
+  7: '88px',
+};
+
+export default spacing;