+++ packages/tooltip/src/Tooltip.tsx
@@ -10,7 +10,7 @@ import Popover, {
 import { useEventListener, useEscapeKey } from '@leafygreen-ui/hooks';
 import { css, cx } from '@leafygreen-ui/emotion';
 import { uiColors } from '@leafygreen-ui/palette';
-import { HTMLElementProps, typeIs } from '@leafygreen-ui/lib';
+import { HTMLElementProps, IdAllocator, typeIs } from '@leafygreen-ui/lib';
 import { transparentize } from 'polished';
 import debounce from 'lodash/debounce';
 import { trianglePosition } from './tooltipUtils';
@@ -176,6 +176,10 @@ export type TooltipProps = Omit<
     }
   >;
 
+const idAllocator = IdAllocator.create('tooltip');
+
+const stopClickPropagation = (evt: React.MouseEvent) => evt.stopPropagation();
+
 /**
  * # Tooltip
  *
@@ -231,10 +235,10 @@ function Tooltip({
 
   const tooltipRef = useRef<HTMLDivElement>(null);
 
-  const tooltipId = useMemo(
-    () => id || `tooltip-${Math.floor(Math.random() * Math.floor(10))}`,
-    [id],
-  );
+  const existingId = id ?? tooltipRef.current?.id;
+  const tooltipId = useMemo(() => existingId ?? idAllocator.generate(), [
+    existingId,
+  ]);
 
   const createTriggerProps = (
     triggerEvent: TriggerEvent,
@@ -308,6 +312,7 @@ function Tooltip({
       justify={justify}
       adjustOnMutation={true}
       spacing={12}
+      onClick={stopClickPropagation}
       {...portalProps}
     >
       {({ align, justify, referenceElPos }: PopoverFunctionParameters) => {
@@ -370,6 +375,7 @@ Tooltip.propTypes = {
   trigger: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
   triggerEvent: PropTypes.oneOf(Object.values(TriggerEvent)),
   variant: PropTypes.oneOf(Object.values(Variant)),
+  enabled: PropTypes.bool,
   open: PropTypes.bool,
   setOpen: PropTypes.func,
   id: PropTypes.string,