+++ packages/typography/src/Link.spec.tsx
@@ -15,8 +15,8 @@ describe('packages/typography', () => {
           href: 'http://mongodb.design',
           arrowAppearance: 'hover',
         });
-        const openInNewTab = screen.getByTitle('Open in New Tab');
-        const rightArrow = screen.queryByTitle('Glyphs / Arrow / Right');
+        const openInNewTab = screen.getByTitle('Open New Tab Icon');
+        const rightArrow = screen.queryByTitle('Arrow Right Icon');
         expect(openInNewTab).toBeInTheDocument();
         expect(rightArrow).not.toBeInTheDocument();
       });
@@ -25,8 +25,8 @@ describe('packages/typography', () => {
         renderLink({
           href: 'http://mongodb.design',
         });
-        const openInNewTab = screen.getByTitle('Open in New Tab');
-        const rightArrow = screen.queryByTitle('Glyphs / Arrow / Right');
+        const openInNewTab = screen.getByTitle('Open New Tab Icon');
+        const rightArrow = screen.queryByTitle('Arrow Right Icon');
         expect(openInNewTab).toBeInTheDocument();
         expect(rightArrow).not.toBeInTheDocument();
       });
@@ -36,7 +36,7 @@ describe('packages/typography', () => {
           hideExternalIcon: true,
           href: 'http://mongodb.design',
         });
-        const openInNewTab = screen.queryByTitle('Open in New Tab');
+        const openInNewTab = screen.queryByTitle('Open New Tab Icon');
         expect(openInNewTab).not.toBeInTheDocument();
       });
     });
@@ -58,8 +58,8 @@ describe('packages/typography', () => {
           arrowAppearance: 'hover',
         });
         const anchor = screen.getByText('Link');
-        const openInNewTab = screen.queryByTitle('Open in New Tab');
-        const rightArrow = screen.queryByTitle('Glyphs / Arrow / Right');
+        const openInNewTab = screen.queryByTitle('Open New Tab Icon');
+        const rightArrow = screen.queryByTitle('Arrow Right Icon');
         expect(rightArrow).not.toBeVisible();
 
         fireEvent.mouseEnter(anchor);
@@ -73,8 +73,8 @@ describe('packages/typography', () => {
           arrowAppearance: 'persist',
         });
 
-        const openInNewTab = screen.queryByTitle('Open in New Tab');
-        const rightArrow = screen.getByTitle('Glyphs / Arrow / Right');
+        const openInNewTab = screen.queryByTitle('Open New Tab Icon');
+        const rightArrow = screen.getByTitle('Arrow Right Icon');
         expect(rightArrow).toBeInTheDocument();
         expect(openInNewTab).not.toBeInTheDocument();
       });
@@ -84,8 +84,8 @@ describe('packages/typography', () => {
           href: 'http://localhost:9001',
         });
         const anchor = screen.getByText('Link');
-        const openInNewTab = screen.queryByTitle('Open in New Tab');
-        const rightArrow = screen.queryByTitle('Glyphs / Arrow / Right');
+        const openInNewTab = screen.queryByTitle('Open New Tab Icon');
+        const rightArrow = screen.queryByTitle('Arrow Right Icon');
 
         fireEvent.mouseEnter(anchor);
         expect(rightArrow).not.toBeInTheDocument();
@@ -97,7 +97,7 @@ describe('packages/typography', () => {
           href: 'http://localhost:9001',
           target: '_blank',
         });
-        const openInNewTab = screen.getByTitle('Open in New Tab');
+        const openInNewTab = screen.getByTitle('Open New Tab Icon');
         expect(openInNewTab).toBeInTheDocument();
       });
     });