+++ packages/stepper/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/stepper
 
+## 1.0.2
+
+### Patch Changes
+
+- 691eb05: Better support for UMD
+- Updated dependencies [691eb05]
+  - @leafygreen-ui/emotion@2.0.2
+  - @leafygreen-ui/icon@6.3.2
+  - @leafygreen-ui/lib@5.1.1
+  - @leafygreen-ui/palette@2.0.2
+
 ## 1.0.1
 
 ### Patch Changes